requirements.txt
datasets>=2.18.0
pandas>=2.2.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn>=0.32.0

